✏️ update : 다른 유저 다이어리 서버 연동 진짜 완료!!!
다른 유저 다이어리 서버 연동 진짜 완료!!!

> Related to : #22
